move I/O and layout in their own modules
